Add files via upload
This program is used to achieve Aggregation in java by using the Shopping parent class, which is then inherited by three child classes Amazon, Flipkart, Mynthra. The google class is used to create the object and the Google_Driver class is the driver class for the Google class, Hence achieving the Aggregation.