Create Google_Driver.java
This program is used to achieve Aggregation in java by using the Shopping class, Which is the parent class from whom the classes Amazon ,Flipkart and Mythra inherit. 
The Google_Driver is the driver class which is used to execute the Google class, And hence creating objects using the methods.